+++ mcp_tools/param_discovery/x8.py
@@ -1,0 +1,38 @@
+# mcp_tools/param_discovery/x8.py
+
+from typing import Dict, Any
+
+def register_x8_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def x8_parameter_discovery(url: str, wordlist: str = "/usr/share/wordlists/x8/params.txt",
+                              method: str = "GET", body: str = "", headers: str = "",
+                              additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute x8 for hidden parameter discovery with enhanced logging.
+
+        Args:
+            url: The target URL
+            wordlist: Parameter wordlist
+            method: HTTP method
+            body: Request body
+            headers: Custom headers
+            additional_args: Additional x8 arguments
+
+        Returns:
+            Hidden parameter discovery results
+        """
+        data = {
+            "url": url,
+            "wordlist": wordlist,
+            "method": method,
+            "body": body,
+            "headers": headers,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting x8 parameter discovery: {url}")
+        result = hexstrike_client.safe_post("api/tools/x8", data)
+        if result.get("success"):
+            logger.info(f"✅ x8 parameter discovery completed for {url}")
+        else:
+            logger.error(f"❌ x8 parameter discovery failed for {url}")
+        return result
