+++ mcp_tools/web_app_security/enhanced_web_app_security.py
@@ -5,74 +5,6 @@ from typing import Dict, Any, Optional
 def register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger, HexStrikeColors):
 
     @mcp.tool()
-    def x8_parameter_discovery(url: str, wordlist: str = "/usr/share/wordlists/x8/params.txt",
-                              method: str = "GET", body: str = "", headers: str = "",
-                              additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute x8 for hidden parameter discovery with enhanced logging.
-
-        Args:
-            url: The target URL
-            wordlist: Parameter wordlist
-            method: HTTP method
-            body: Request body
-            headers: Custom headers
-            additional_args: Additional x8 arguments
-
-        Returns:
-            Hidden parameter discovery results
-        """
-        data = {
-            "url": url,
-            "wordlist": wordlist,
-            "method": method,
-            "body": body,
-            "headers": headers,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting x8 parameter discovery: {url}")
-        result = hexstrike_client.safe_post("api/tools/x8", data)
-        if result.get("success"):
-            logger.info(f"✅ x8 parameter discovery completed for {url}")
-        else:
-            logger.error(f"❌ x8 parameter discovery failed for {url}")
-        return result
-
-    @mcp.tool()
-    def jaeles_vulnerability_scan(url: str, signatures: str = "", config: str = "",
-                                 threads: int = 20, timeout: int = 20,
-                                 additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Jaeles for advanced vulnerability scanning with custom signatures.
-
-        Args:
-            url: The target URL
-            signatures: Custom signature path
-            config: Configuration file
-            threads: Number of threads
-            timeout: Request timeout
-            additional_args: Additional Jaeles arguments
-
-        Returns:
-            Advanced vulnerability scanning results with custom signatures
-        """
-        data = {
-            "url": url,
-            "signatures": signatures,
-            "config": config,
-            "threads": threads,
-            "timeout": timeout,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔬 Starting Jaeles vulnerability scan: {url}")
-        result = hexstrike_client.safe_post("api/tools/jaeles", data)
-        if result.get("success"):
-            logger.info(f"✅ Jaeles vulnerability scan completed for {url}")
-        else:
-            logger.error(f"❌ Jaeles vulnerability scan failed for {url}")
-        return result
-
-    @mcp.tool()
     def dalfox_xss_scan(url: str, pipe_mode: bool = False, blind: bool = False,
                        mining_dom: bool = True, mining_dict: bool = True,
                        custom_payload: str = "", additional_args: str = "") -> Dict[str, Any]:
