+++ mcp_tools/web_scan/jaeles.py
@@ -1,0 +1,38 @@
+# mcp_tools/web_scan/jaeles.py
+
+from typing import Dict, Any
+
+def register_jaeles_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def jaeles_vulnerability_scan(url: str, signatures: str = "", config: str = "",
+                                 threads: int = 20, timeout: int = 20,
+                                 additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Jaeles for advanced vulnerability scanning with custom signatures.
+
+        Args:
+            url: The target URL
+            signatures: Custom signature path
+            config: Configuration file
+            threads: Number of threads
+            timeout: Request timeout
+            additional_args: Additional Jaeles arguments
+
+        Returns:
+            Advanced vulnerability scanning results with custom signatures
+        """
+        data = {
+            "url": url,
+            "signatures": signatures,
+            "config": config,
+            "threads": threads,
+            "timeout": timeout,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔬 Starting Jaeles vulnerability scan: {url}")
+        result = hexstrike_client.safe_post("api/tools/jaeles", data)
+        if result.get("success"):
+            logger.info(f"✅ Jaeles vulnerability scan completed for {url}")
+        else:
+            logger.error(f"❌ Jaeles vulnerability scan failed for {url}")
+        return result
